chain: Cleanup consensus logic
