+++ chain/consensus/filcns/mine.go
@@ -1,38 +1,35 @@
-package gen
+package filcns
 
 import (
 	"context"
+	"github.com/filecoin-project/lotus/chain/consensus"
 
-	"github.com/filecoin-project/go-state-types/crypto"
-	blockadt "github.com/filecoin-project/specs-actors/actors/util/adt"
-	cid "github.com/ipfs/go-cid"
-	cbg "github.com/whyrusleeping/cbor-gen"
+	"github.com/ipfs/go-cid"
 	"golang.org/x/xerrors"
 
-	ffi "github.com/filecoin-project/filecoin-ffi"
+	"github.com/filecoin-project/go-state-types/crypto"
 	"github.com/filecoin-project/lotus/api"
 	"github.com/filecoin-project/lotus/chain/stmgr"
 	"github.com/filecoin-project/lotus/chain/types"
 )
 
-func MinerCreateBlock(ctx context.Context, sm *stmgr.StateManager, w api.Wallet, bt *api.BlockTemplate) (*types.FullBlock, error) {
-
-	pts, err := sm.ChainStore().LoadTipSet(bt.Parents)
+func (filec *FilecoinEC) CreateBlock(ctx context.Context, w api.Wallet, bt *api.BlockTemplate) (*types.FullBlock, error) {
+	pts, err := filec.sm.ChainStore().LoadTipSet(bt.Parents)
 	if err != nil {
 		return nil, xerrors.Errorf("failed to load parent tipset: %w", err)
 	}
 
-	st, recpts, err := sm.TipSetState(ctx, pts)
+	st, recpts, err := filec.sm.TipSetState(ctx, pts)
 	if err != nil {
 		return nil, xerrors.Errorf("failed to load tipset state: %w", err)
 	}
 
-	_, lbst, err := stmgr.GetLookbackTipSetForRound(ctx, sm, pts, bt.Epoch)
+	_, lbst, err := stmgr.GetLookbackTipSetForRound(ctx, filec.sm, pts, bt.Epoch)
 	if err != nil {
 		return nil, xerrors.Errorf("getting lookback miner actor state: %w", err)
 	}
 
-	worker, err := stmgr.GetMinerWorkerRaw(ctx, sm, lbst, bt.Miner)
+	worker, err := stmgr.GetMinerWorkerRaw(ctx, filec.sm, lbst, bt.Miner)
 	if err != nil {
 		return nil, xerrors.Errorf("failed to get miner worker: %w", err)
 	}
@@ -61,14 +58,14 @@ func MinerCreateBlock(ctx context.Context, sm *stmgr.StateManager, w api.Wallet,
 			blsSigs = append(blsSigs, msg.Signature)
 			blsMessages = append(blsMessages, &msg.Message)
 
-			c, err := sm.ChainStore().PutMessage(&msg.Message)
+			c, err := filec.sm.ChainStore().PutMessage(&msg.Message)
 			if err != nil {
 				return nil, err
 			}
 
 			blsMsgCids = append(blsMsgCids, c)
 		} else {
-			c, err := sm.ChainStore().PutMessage(msg)
+			c, err := filec.sm.ChainStore().PutMessage(msg)
 			if err != nil {
 				return nil, err
 			}
@@ -79,12 +76,12 @@ func MinerCreateBlock(ctx context.Context, sm *stmgr.StateManager, w api.Wallet,
 		}
 	}
 
-	store := sm.ChainStore().ActorStore(ctx)
-	blsmsgroot, err := toArray(store, blsMsgCids)
+	store := filec.sm.ChainStore().ActorStore(ctx)
+	blsmsgroot, err := consensus.ToMessagesArray(store, blsMsgCids)
 	if err != nil {
 		return nil, xerrors.Errorf("building bls amt: %w", err)
 	}
-	secpkmsgroot, err := toArray(store, secpkMsgCids)
+	secpkmsgroot, err := consensus.ToMessagesArray(store, secpkMsgCids)
 	if err != nil {
 		return nil, xerrors.Errorf("building secpk amt: %w", err)
 	}
@@ -98,19 +95,19 @@ func MinerCreateBlock(ctx context.Context, sm *stmgr.StateManager, w api.Wallet,
 	}
 	next.Messages = mmcid
 
-	aggSig, err := aggregateSignatures(blsSigs)
+	aggSig, err := consensus.AggregateSignatures(blsSigs)
 	if err != nil {
 		return nil, err
 	}
 
 	next.BLSAggregate = aggSig
-	pweight, err := sm.ChainStore().Weight(ctx, pts)
+	pweight, err := filec.sm.ChainStore().Weight(ctx, pts)
 	if err != nil {
 		return nil, err
 	}
 	next.ParentWeight = pweight
 
-	baseFee, err := sm.ChainStore().ComputeBaseFee(ctx, pts)
+	baseFee, err := filec.sm.ChainStore().ComputeBaseFee(ctx, pts)
 	if err != nil {
 		return nil, xerrors.Errorf("computing base fee: %w", err)
 	}
@@ -137,42 +134,4 @@ func MinerCreateBlock(ctx context.Context, sm *stmgr.StateManager, w api.Wallet,
 	}
 
 	return fullBlock, nil
-}
-
-func aggregateSignatures(sigs []crypto.Signature) (*crypto.Signature, error) {
-	sigsS := make([]ffi.Signature, len(sigs))
-	for i := 0; i < len(sigs); i++ {
-		copy(sigsS[i][:], sigs[i].Data[:ffi.SignatureBytes])
-	}
-
-	aggSig := ffi.Aggregate(sigsS)
-	if aggSig == nil {
-		if len(sigs) > 0 {
-			return nil, xerrors.Errorf("bls.Aggregate returned nil with %d signatures", len(sigs))
-		}
-
-		zeroSig := ffi.CreateZeroSignature()
-
-		// Note: for blst this condition should not happen - nil should not
-		// be returned
-		return &crypto.Signature{
-			Type: crypto.SigTypeBLS,
-			Data: zeroSig[:],
-		}, nil
-	}
-	return &crypto.Signature{
-		Type: crypto.SigTypeBLS,
-		Data: aggSig[:],
-	}, nil
-}
-
-func toArray(store blockadt.Store, cids []cid.Cid) (cid.Cid, error) {
-	arr := blockadt.MakeEmptyArray(store)
-	for i, c := range cids {
-		oc := cbg.CborCid(c)
-		if err := arr.Set(uint64(i), &oc); err != nil {
-			return cid.Undef, err
-		}
-	}
-	return arr.Root()
 }
