+++ chain/stmgr/forks.go
@@ -15,8 +15,6 @@ import (
 	"github.com/filecoin-project/go-state-types/abi"
 	"github.com/filecoin-project/go-state-types/big"
 	"github.com/filecoin-project/go-state-types/network"
-	"github.com/filecoin-project/go-state-types/rt"
-
 	"github.com/filecoin-project/specs-actors/v3/actors/migration/nv10"
 
 	"github.com/filecoin-project/lotus/chain/actors/adt"
@@ -98,21 +96,6 @@ type Upgrade struct {
 
 type UpgradeSchedule []Upgrade
 
-type migrationLogger struct{}
-
-func (ml migrationLogger) Log(level rt.LogLevel, msg string, args ...interface{}) {
-	switch level {
-	case rt.DEBUG:
-		log.Debugf(msg, args...)
-	case rt.INFO:
-		log.Infof(msg, args...)
-	case rt.WARN:
-		log.Warnf(msg, args...)
-	case rt.ERROR:
-		log.Errorf(msg, args...)
-	}
-}
-
 func (us UpgradeSchedule) Validate() error {
 	// Make sure each upgrade is valid.
 	for _, u := range us {
@@ -181,7 +164,7 @@ func (us UpgradeSchedule) GetNtwkVersion(e abi.ChainEpoch) (network.Version, err
 	return network.Version0, xerrors.Errorf("Epoch %d has no defined network version", e)
 }
 
-func (sm *StateManager) handleStateForks(ctx context.Context, root cid.Cid, height abi.ChainEpoch, cb ExecMonitor, ts *types.TipSet) (cid.Cid, error) {
+func (sm *StateManager) HandleStateForks(ctx context.Context, root cid.Cid, height abi.ChainEpoch, cb ExecMonitor, ts *types.TipSet) (cid.Cid, error) {
 	retCid := root
 	var err error
 	u := sm.stateMigrations[height]
@@ -331,7 +314,7 @@ func (sm *StateManager) preMigrationWorker(ctx context.Context) {
 	}
 }
 
-func doTransfer(tree types.StateTree, from, to address.Address, amt abi.TokenAmount, cb func(trace types.ExecutionTrace)) error {
+func DoTransfer(tree types.StateTree, from, to address.Address, amt abi.TokenAmount, cb func(trace types.ExecutionTrace)) error {
 	fromAct, err := tree.GetActor(from)
 	if err != nil {
 		return xerrors.Errorf("failed to get 'from' actor for transfer: %w", err)
@@ -361,8 +344,8 @@ func doTransfer(tree types.StateTree, from, to address.Address, amt abi.TokenAmo
 		// record the transfer in execution traces
 
 		cb(types.ExecutionTrace{
-			Msg:        makeFakeMsg(from, to, amt, 0),
-			MsgRct:     makeFakeRct(),
+			Msg:        MakeFakeMsg(from, to, amt, 0),
+			MsgRct:     MakeFakeRct(),
 			Error:      "",
 			Duration:   0,
 			GasCharges: nil,
@@ -373,7 +356,7 @@ func doTransfer(tree types.StateTree, from, to address.Address, amt abi.TokenAmo
 	return nil
 }
 
-func terminateActor(ctx context.Context, tree *state.StateTree, addr address.Address, em ExecMonitor, epoch abi.ChainEpoch, ts *types.TipSet) error {
+func TerminateActor(ctx context.Context, tree *state.StateTree, addr address.Address, em ExecMonitor, epoch abi.ChainEpoch, ts *types.TipSet) error {
 	a, err := tree.GetActor(addr)
 	if xerrors.Is(err, types.ErrActorNotFound) {
 		return types.ErrActorNotFound
@@ -382,7 +365,7 @@ func terminateActor(ctx context.Context, tree *state.StateTree, addr address.Add
 	}
 
 	var trace types.ExecutionTrace
-	if err := doTransfer(tree, addr, builtin.BurntFundsActorAddr, a.Balance, func(t types.ExecutionTrace) {
+	if err := DoTransfer(tree, addr, builtin.BurntFundsActorAddr, a.Balance, func(t types.ExecutionTrace) {
 		trace = t
 	}); err != nil {
 		return xerrors.Errorf("transferring terminated actor's balance: %w", err)
@@ -391,10 +374,10 @@ func terminateActor(ctx context.Context, tree *state.StateTree, addr address.Add
 	if em != nil {
 		// record the transfer in execution traces
 
-		fakeMsg := makeFakeMsg(builtin.SystemActorAddr, addr, big.Zero(), uint64(epoch))
+		fakeMsg := MakeFakeMsg(builtin.SystemActorAddr, addr, big.Zero(), uint64(epoch))
 
 		if err := em.MessageApplied(ctx, ts, fakeMsg.Cid(), fakeMsg, &vm.ApplyRet{
-			MessageReceipt: *makeFakeRct(),
+			MessageReceipt: *MakeFakeRct(),
 			ActorErr:       nil,
 			ExecutionTrace: trace,
 			Duration:       0,
@@ -433,7 +416,7 @@ func terminateActor(ctx context.Context, tree *state.StateTree, addr address.Add
 	return tree.SetActor(init_.Address, ia)
 }
 
-func setNetworkName(ctx context.Context, store adt.Store, tree *state.StateTree, name string) error {
+func SetNetworkName(ctx context.Context, store adt.Store, tree *state.StateTree, name string) error {
 	ia, err := tree.GetActor(init_.Address)
 	if err != nil {
 		return xerrors.Errorf("getting init actor: %w", err)
@@ -460,7 +443,7 @@ func setNetworkName(ctx context.Context, store adt.Store, tree *state.StateTree,
 	return nil
 }
 
-func makeKeyAddr(splitAddr address.Address, count uint64) (address.Address, error) {
+func MakeKeyAddr(splitAddr address.Address, count uint64) (address.Address, error) {
 	var b bytes.Buffer
 	if err := splitAddr.MarshalCBOR(&b); err != nil {
 		return address.Undef, xerrors.Errorf("marshalling split address: %w", err)
@@ -482,7 +465,7 @@ func makeKeyAddr(splitAddr address.Address, count uint64) (address.Address, erro
 	return addr, nil
 }
 
-func makeFakeMsg(from address.Address, to address.Address, amt abi.TokenAmount, nonce uint64) *types.Message {
+func MakeFakeMsg(from address.Address, to address.Address, amt abi.TokenAmount, nonce uint64) *types.Message {
 	return &types.Message{
 		From:  from,
 		To:    to,
@@ -491,7 +474,7 @@ func makeFakeMsg(from address.Address, to address.Address, amt abi.TokenAmount, 
 	}
 }
 
-func makeFakeRct() *types.MessageReceipt {
+func MakeFakeRct() *types.MessageReceipt {
 	return &types.MessageReceipt{
 		ExitCode: 0,
 		Return:   nil,
