+++ chain/sync.go
@@ -5,13 +5,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"os"
 	"sort"
-	"strings"
 	"sync"
 	"time"
 
-	"github.com/filecoin-project/lotus/chain/actors/builtin"
+	"github.com/filecoin-project/lotus/chain/consensus"
 
 	"github.com/filecoin-project/lotus/node/modules/dtypes"
 
@@ -29,39 +27,22 @@ import (
 	"go.opencensus.io/trace"
 	"golang.org/x/xerrors"
 
-	"github.com/filecoin-project/go-address"
-	"github.com/filecoin-project/go-state-types/abi"
 	"github.com/filecoin-project/go-state-types/crypto"
-	"github.com/filecoin-project/go-state-types/network"
-	"github.com/filecoin-project/lotus/extern/sector-storage/ffiwrapper"
-
-	ffi "github.com/filecoin-project/filecoin-ffi"
-
 	// named msgarray here to make it clear that these are the types used by
 	// messages, regardless of specs-actors version.
 	blockadt "github.com/filecoin-project/specs-actors/actors/util/adt"
 
-	proof2 "github.com/filecoin-project/specs-actors/v2/actors/runtime/proof"
-
 	"github.com/filecoin-project/lotus/api"
 	bstore "github.com/filecoin-project/lotus/blockstore"
 	"github.com/filecoin-project/lotus/build"
-	"github.com/filecoin-project/lotus/chain/actors/builtin/power"
 	"github.com/filecoin-project/lotus/chain/beacon"
 	"github.com/filecoin-project/lotus/chain/exchange"
-	"github.com/filecoin-project/lotus/chain/gen"
-	"github.com/filecoin-project/lotus/chain/state"
 	"github.com/filecoin-project/lotus/chain/stmgr"
 	"github.com/filecoin-project/lotus/chain/store"
 	"github.com/filecoin-project/lotus/chain/types"
 	"github.com/filecoin-project/lotus/chain/vm"
-	"github.com/filecoin-project/lotus/lib/sigs"
 	"github.com/filecoin-project/lotus/metrics"
 )
-
-// Blocks that are more than MaxHeightDrift epochs above
-// the theoretical max height based on systime are quickly rejected
-const MaxHeightDrift = 5
 
 var (
 	// LocalIncoming is the _local_ pubsub (unrelated to libp2p pubsub) topic
@@ -108,6 +89,8 @@ type Syncer struct {
 	// the state manager handles making state queries
 	sm *stmgr.StateManager
 
+	consensus consensus.Consensus
+
 	// The known Genesis tipset
 	Genesis *types.TipSet
 
@@ -127,8 +110,6 @@ type Syncer struct {
 
 	receiptTracker *blockReceiptTracker
 
-	verifier ffiwrapper.Verifier
-
 	tickerCtxCancel context.CancelFunc
 
 	ds dtypes.MetadataDS
@@ -136,38 +117,42 @@ type Syncer struct {
 
 type SyncManagerCtor func(syncFn SyncFunc) SyncManager
 
-// NewSyncer creates a new Syncer object.
-func NewSyncer(ds dtypes.MetadataDS, sm *stmgr.StateManager, exchange exchange.Client, syncMgrCtor SyncManagerCtor, connmgr connmgr.ConnManager, self peer.ID, beacon beacon.Schedule, verifier ffiwrapper.Verifier) (*Syncer, error) {
+type Genesis *types.TipSet
+
+func LoadGenesis(sm *stmgr.StateManager) (Genesis, error) {
 	gen, err := sm.ChainStore().GetGenesis()
 	if err != nil {
 		return nil, xerrors.Errorf("getting genesis block: %w", err)
 	}
 
-	gent, err := types.NewTipSet([]*types.BlockHeader{gen})
-	if err != nil {
-		return nil, err
-	}
+	return types.NewTipSet([]*types.BlockHeader{gen})
+}
+
+// NewSyncer creates a new Syncer object.
+func NewSyncer(ds dtypes.MetadataDS,
+	sm *stmgr.StateManager,
+	exchange exchange.Client,
+	syncMgrCtor SyncManagerCtor,
+	connmgr connmgr.ConnManager,
+	self peer.ID,
+	beacon beacon.Schedule,
+	gent Genesis,
+	consensus consensus.Consensus) (*Syncer, error) {
 
 	s := &Syncer{
 		ds:             ds,
 		beacon:         beacon,
 		bad:            NewBadBlockCache(),
 		Genesis:        gent,
+		consensus:      consensus,
 		Exchange:       exchange,
 		store:          sm.ChainStore(),
 		sm:             sm,
 		self:           self,
 		receiptTracker: newBlockReceiptTracker(),
 		connmgr:        connmgr,
-		verifier:       verifier,
 
 		incoming: pubsub.New(50),
-	}
-
-	if build.InsecurePoStValidation {
-		log.Warn("*********************************************************************************************")
-		log.Warn(" [INSECURE-POST-VALIDATION] Insecure test validation is enabled. If you see this outside of a test, it is a severe bug! ")
-		log.Warn("*********************************************************************************************")
 	}
 
 	s.syncmgr = syncMgrCtor(s.Sync)
@@ -212,7 +197,7 @@ func (syncer *Syncer) Stop() {
 func (syncer *Syncer) InformNewHead(from peer.ID, fts *store.FullTipSet) bool {
 	defer func() {
 		if err := recover(); err != nil {
-			log.Errorf("panic in InformNewHead: ", err)
+			log.Errorf("panic in InformNewHead: %s", err)
 		}
 	}()
 
@@ -222,7 +207,7 @@ func (syncer *Syncer) InformNewHead(from peer.ID, fts *store.FullTipSet) bool {
 		return false
 	}
 
-	if syncer.IsEpochBeyondCurrMax(fts.TipSet().Height()) {
+	if syncer.consensus.IsEpochBeyondCurrMax(fts.TipSet().Height()) {
 		log.Errorf("Received block with impossibly large height %d", fts.TipSet().Height())
 		return false
 	}
@@ -399,33 +384,21 @@ func zipTipSetAndMessages(bs cbor.IpldStore, ts *types.TipSet, allbmsgs []*types
 		return nil, fmt.Errorf("msgincl length didnt match tipset size")
 	}
 
+	if err := checkMsgMeta(ts, allbmsgs, allsmsgs, bmi, smi); err != nil {
+		return nil, err
+	}
+
 	fts := &store.FullTipSet{}
 	for bi, b := range ts.Blocks() {
-		if msgc := len(bmi[bi]) + len(smi[bi]); msgc > build.BlockMessageLimit {
-			return nil, fmt.Errorf("block %q has too many messages (%d)", b.Cid(), msgc)
-		}
 
 		var smsgs []*types.SignedMessage
-		var smsgCids []cid.Cid
 		for _, m := range smi[bi] {
 			smsgs = append(smsgs, allsmsgs[m])
-			smsgCids = append(smsgCids, allsmsgs[m].Cid())
 		}
 
 		var bmsgs []*types.Message
-		var bmsgCids []cid.Cid
 		for _, m := range bmi[bi] {
 			bmsgs = append(bmsgs, allbmsgs[m])
-			bmsgCids = append(bmsgCids, allbmsgs[m].Cid())
-		}
-
-		mrcid, err := computeMsgMeta(bs, bmsgCids, smsgCids)
-		if err != nil {
-			return nil, err
-		}
-
-		if b.Messages != mrcid {
-			return nil, fmt.Errorf("messages didnt match message root in header for ts %s", ts.Key())
 		}
 
 		fb := &types.FullBlock{
@@ -584,7 +557,7 @@ func (syncer *Syncer) Sync(ctx context.Context, maybeHead *types.TipSet) error {
 }
 
 func isPermanent(err error) bool {
-	return !errors.Is(err, ErrTemporal)
+	return !errors.Is(err, consensus.ErrTemporal)
 }
 
 func (syncer *Syncer) ValidateTipSet(ctx context.Context, fts *store.FullTipSet, useCache bool) error {
@@ -624,55 +597,6 @@ func (syncer *Syncer) ValidateTipSet(ctx context.Context, fts *store.FullTipSet,
 	return nil
 }
 
-func (syncer *Syncer) minerIsValid(ctx context.Context, maddr address.Address, baseTs *types.TipSet) error {
-	act, err := syncer.sm.LoadActor(ctx, power.Address, baseTs)
-	if err != nil {
-		return xerrors.Errorf("failed to load power actor: %w", err)
-	}
-
-	powState, err := power.Load(syncer.store.ActorStore(ctx), act)
-	if err != nil {
-		return xerrors.Errorf("failed to load power actor state: %w", err)
-	}
-
-	_, exist, err := powState.MinerPower(maddr)
-	if err != nil {
-		return xerrors.Errorf("failed to look up miner's claim: %w", err)
-	}
-
-	if !exist {
-		return xerrors.New("miner isn't valid")
-	}
-
-	return nil
-}
-
-var ErrTemporal = errors.New("temporal error")
-
-func blockSanityChecks(h *types.BlockHeader) error {
-	if h.ElectionProof == nil {
-		return xerrors.Errorf("block cannot have nil election proof")
-	}
-
-	if h.Ticket == nil {
-		return xerrors.Errorf("block cannot have nil ticket")
-	}
-
-	if h.BlockSig == nil {
-		return xerrors.Errorf("block had nil signature")
-	}
-
-	if h.BLSAggregate == nil {
-		return xerrors.Errorf("block had nil bls aggregate signature")
-	}
-
-	if h.Miner.Protocol() != address.ID {
-		return xerrors.Errorf("block had non-ID miner address")
-	}
-
-	return nil
-}
-
 // ValidateBlock should match up with 'Semantical Validation' in validation.md in the spec
 func (syncer *Syncer) ValidateBlock(ctx context.Context, b *types.FullBlock, useCache bool) (err error) {
 	defer func() {
@@ -703,262 +627,8 @@ func (syncer *Syncer) ValidateBlock(ctx context.Context, b *types.FullBlock, use
 	ctx, span := trace.StartSpan(ctx, "validateBlock")
 	defer span.End()
 
-	if err := blockSanityChecks(b.Header); err != nil {
-		return xerrors.Errorf("incoming header failed basic sanity checks: %w", err)
-	}
-
-	h := b.Header
-
-	baseTs, err := syncer.store.LoadTipSet(types.NewTipSetKey(h.Parents...))
-	if err != nil {
-		return xerrors.Errorf("load parent tipset failed (%s): %w", h.Parents, err)
-	}
-
-	winPoStNv := syncer.sm.GetNtwkVersion(ctx, baseTs.Height())
-
-	lbts, lbst, err := stmgr.GetLookbackTipSetForRound(ctx, syncer.sm, baseTs, h.Height)
-	if err != nil {
-		return xerrors.Errorf("failed to get lookback tipset for block: %w", err)
-	}
-
-	prevBeacon, err := syncer.store.GetLatestBeaconEntry(baseTs)
-	if err != nil {
-		return xerrors.Errorf("failed to get latest beacon entry: %w", err)
-	}
-
-	// fast checks first
-
-	if h.Height <= baseTs.Height() {
-		return xerrors.Errorf("block height not greater than parent height: %d != %d", h.Height, baseTs.Height())
-	}
-
-	nulls := h.Height - (baseTs.Height() + 1)
-	if tgtTs := baseTs.MinTimestamp() + build.BlockDelaySecs*uint64(nulls+1); h.Timestamp != tgtTs {
-		return xerrors.Errorf("block has wrong timestamp: %d != %d", h.Timestamp, tgtTs)
-	}
-
-	now := uint64(build.Clock.Now().Unix())
-	if h.Timestamp > now+build.AllowableClockDriftSecs {
-		return xerrors.Errorf("block was from the future (now=%d, blk=%d): %w", now, h.Timestamp, ErrTemporal)
-	}
-	if h.Timestamp > now {
-		log.Warn("Got block from the future, but within threshold", h.Timestamp, build.Clock.Now().Unix())
-	}
-
-	msgsCheck := async.Err(func() error {
-		if b.Cid() == build.WhitelistedBlock {
-			return nil
-		}
-
-		if err := syncer.checkBlockMessages(ctx, b, baseTs); err != nil {
-			return xerrors.Errorf("block had invalid messages: %w", err)
-		}
-		return nil
-	})
-
-	minerCheck := async.Err(func() error {
-		if err := syncer.minerIsValid(ctx, h.Miner, baseTs); err != nil {
-			return xerrors.Errorf("minerIsValid failed: %w", err)
-		}
-		return nil
-	})
-
-	baseFeeCheck := async.Err(func() error {
-		baseFee, err := syncer.store.ComputeBaseFee(ctx, baseTs)
-		if err != nil {
-			return xerrors.Errorf("computing base fee: %w", err)
-		}
-		if types.BigCmp(baseFee, b.Header.ParentBaseFee) != 0 {
-			return xerrors.Errorf("base fee doesn't match: %s (header) != %s (computed)",
-				b.Header.ParentBaseFee, baseFee)
-		}
-		return nil
-	})
-	pweight, err := syncer.store.Weight(ctx, baseTs)
-	if err != nil {
-		return xerrors.Errorf("getting parent weight: %w", err)
-	}
-
-	if types.BigCmp(pweight, b.Header.ParentWeight) != 0 {
-		return xerrors.Errorf("parrent weight different: %s (header) != %s (computed)",
-			b.Header.ParentWeight, pweight)
-	}
-
-	stateRootCheck := async.Err(func() error {
-		stateroot, precp, err := syncer.sm.TipSetState(ctx, baseTs)
-		if err != nil {
-			return xerrors.Errorf("get tipsetstate(%d, %s) failed: %w", h.Height, h.Parents, err)
-		}
-
-		if stateroot != h.ParentStateRoot {
-			msgs, err := syncer.store.MessagesForTipset(baseTs)
-			if err != nil {
-				log.Error("failed to load messages for tipset during tipset state mismatch error: ", err)
-			} else {
-				log.Warn("Messages for tipset with mismatching state:")
-				for i, m := range msgs {
-					mm := m.VMMessage()
-					log.Warnf("Message[%d]: from=%s to=%s method=%d params=%x", i, mm.From, mm.To, mm.Method, mm.Params)
-				}
-			}
-
-			return xerrors.Errorf("parent state root did not match computed state (%s != %s)", stateroot, h.ParentStateRoot)
-		}
-
-		if precp != h.ParentMessageReceipts {
-			return xerrors.Errorf("parent receipts root did not match computed value (%s != %s)", precp, h.ParentMessageReceipts)
-		}
-
-		return nil
-	})
-
-	// Stuff that needs worker address
-	waddr, err := stmgr.GetMinerWorkerRaw(ctx, syncer.sm, lbst, h.Miner)
-	if err != nil {
-		return xerrors.Errorf("GetMinerWorkerRaw failed: %w", err)
-	}
-
-	winnerCheck := async.Err(func() error {
-		if h.ElectionProof.WinCount < 1 {
-			return xerrors.Errorf("block is not claiming to be a winner")
-		}
-
-		eligible, err := stmgr.MinerEligibleToMine(ctx, syncer.sm, h.Miner, baseTs, lbts)
-		if err != nil {
-			return xerrors.Errorf("determining if miner has min power failed: %w", err)
-		}
-
-		if !eligible {
-			return xerrors.New("block's miner is ineligible to mine")
-		}
-
-		rBeacon := *prevBeacon
-		if len(h.BeaconEntries) != 0 {
-			rBeacon = h.BeaconEntries[len(h.BeaconEntries)-1]
-		}
-		buf := new(bytes.Buffer)
-		if err := h.Miner.MarshalCBOR(buf); err != nil {
-			return xerrors.Errorf("failed to marshal miner address to cbor: %w", err)
-		}
-
-		vrfBase, err := store.DrawRandomness(rBeacon.Data, crypto.DomainSeparationTag_ElectionProofProduction, h.Height, buf.Bytes())
-		if err != nil {
-			return xerrors.Errorf("could not draw randomness: %w", err)
-		}
-
-		if err := VerifyElectionPoStVRF(ctx, waddr, vrfBase, h.ElectionProof.VRFProof); err != nil {
-			return xerrors.Errorf("validating block election proof failed: %w", err)
-		}
-
-		slashed, err := stmgr.GetMinerSlashed(ctx, syncer.sm, baseTs, h.Miner)
-		if err != nil {
-			return xerrors.Errorf("failed to check if block miner was slashed: %w", err)
-		}
-
-		if slashed {
-			return xerrors.Errorf("received block was from slashed or invalid miner")
-		}
-
-		mpow, tpow, _, err := stmgr.GetPowerRaw(ctx, syncer.sm, lbst, h.Miner)
-		if err != nil {
-			return xerrors.Errorf("failed getting power: %w", err)
-		}
-
-		j := h.ElectionProof.ComputeWinCount(mpow.QualityAdjPower, tpow.QualityAdjPower)
-		if h.ElectionProof.WinCount != j {
-			return xerrors.Errorf("miner claims wrong number of wins: miner: %d, computed: %d", h.ElectionProof.WinCount, j)
-		}
-
-		return nil
-	})
-
-	blockSigCheck := async.Err(func() error {
-		if err := sigs.CheckBlockSignature(ctx, h, waddr); err != nil {
-			return xerrors.Errorf("check block signature failed: %w", err)
-		}
-		return nil
-	})
-
-	beaconValuesCheck := async.Err(func() error {
-		if os.Getenv("LOTUS_IGNORE_DRAND") == "_yes_" {
-			return nil
-		}
-
-		if err := beacon.ValidateBlockValues(syncer.beacon, h, baseTs.Height(), *prevBeacon); err != nil {
-			return xerrors.Errorf("failed to validate blocks random beacon values: %w", err)
-		}
-		return nil
-	})
-
-	tktsCheck := async.Err(func() error {
-		buf := new(bytes.Buffer)
-		if err := h.Miner.MarshalCBOR(buf); err != nil {
-			return xerrors.Errorf("failed to marshal miner address to cbor: %w", err)
-		}
-
-		if h.Height > build.UpgradeSmokeHeight {
-			buf.Write(baseTs.MinTicket().VRFProof)
-		}
-
-		beaconBase := *prevBeacon
-		if len(h.BeaconEntries) != 0 {
-			beaconBase = h.BeaconEntries[len(h.BeaconEntries)-1]
-		}
-
-		vrfBase, err := store.DrawRandomness(beaconBase.Data, crypto.DomainSeparationTag_TicketProduction, h.Height-build.TicketRandomnessLookback, buf.Bytes())
-		if err != nil {
-			return xerrors.Errorf("failed to compute vrf base for ticket: %w", err)
-		}
-
-		err = VerifyElectionPoStVRF(ctx, waddr, vrfBase, h.Ticket.VRFProof)
-		if err != nil {
-			return xerrors.Errorf("validating block tickets failed: %w", err)
-		}
-		return nil
-	})
-
-	wproofCheck := async.Err(func() error {
-		if err := syncer.VerifyWinningPoStProof(ctx, winPoStNv, h, *prevBeacon, lbst, waddr); err != nil {
-			return xerrors.Errorf("invalid election post: %w", err)
-		}
-		return nil
-	})
-
-	await := []async.ErrorFuture{
-		minerCheck,
-		tktsCheck,
-		blockSigCheck,
-		beaconValuesCheck,
-		wproofCheck,
-		winnerCheck,
-		msgsCheck,
-		baseFeeCheck,
-		stateRootCheck,
-	}
-
-	var merr error
-	for _, fut := range await {
-		if err := fut.AwaitContext(ctx); err != nil {
-			merr = multierror.Append(merr, err)
-		}
-	}
-	if merr != nil {
-		mulErr := merr.(*multierror.Error)
-		mulErr.ErrorFormat = func(es []error) string {
-			if len(es) == 1 {
-				return fmt.Sprintf("1 error occurred:\n\t* %+v\n\n", es[0])
-			}
-
-			points := make([]string, len(es))
-			for i, err := range es {
-				points[i] = fmt.Sprintf("* %+v", err)
-			}
-
-			return fmt.Sprintf(
-				"%d errors occurred:\n\t%s\n\n",
-				len(es), strings.Join(points, "\n\t"))
-		}
-		return mulErr
+	if err := syncer.consensus.ValidateBlock(ctx, b); err != nil {
+		return err
 	}
 
 	if useCache {
@@ -967,249 +637,6 @@ func (syncer *Syncer) ValidateBlock(ctx context.Context, b *types.FullBlock, use
 		}
 	}
 
-	return nil
-}
-
-func (syncer *Syncer) VerifyWinningPoStProof(ctx context.Context, nv network.Version, h *types.BlockHeader, prevBeacon types.BeaconEntry, lbst cid.Cid, waddr address.Address) error {
-	if build.InsecurePoStValidation {
-		if len(h.WinPoStProof) == 0 {
-			return xerrors.Errorf("[INSECURE-POST-VALIDATION] No winning post proof given")
-		}
-
-		if string(h.WinPoStProof[0].ProofBytes) == "valid proof" {
-			return nil
-		}
-		return xerrors.Errorf("[INSECURE-POST-VALIDATION] winning post was invalid")
-	}
-
-	buf := new(bytes.Buffer)
-	if err := h.Miner.MarshalCBOR(buf); err != nil {
-		return xerrors.Errorf("failed to marshal miner address: %w", err)
-	}
-
-	rbase := prevBeacon
-	if len(h.BeaconEntries) > 0 {
-		rbase = h.BeaconEntries[len(h.BeaconEntries)-1]
-	}
-
-	rand, err := store.DrawRandomness(rbase.Data, crypto.DomainSeparationTag_WinningPoStChallengeSeed, h.Height, buf.Bytes())
-	if err != nil {
-		return xerrors.Errorf("failed to get randomness for verifying winning post proof: %w", err)
-	}
-
-	mid, err := address.IDFromAddress(h.Miner)
-	if err != nil {
-		return xerrors.Errorf("failed to get ID from miner address %s: %w", h.Miner, err)
-	}
-
-	sectors, err := stmgr.GetSectorsForWinningPoSt(ctx, nv, syncer.verifier, syncer.sm, lbst, h.Miner, rand)
-	if err != nil {
-		return xerrors.Errorf("getting winning post sector set: %w", err)
-	}
-
-	ok, err := ffiwrapper.ProofVerifier.VerifyWinningPoSt(ctx, proof2.WinningPoStVerifyInfo{
-		Randomness:        rand,
-		Proofs:            h.WinPoStProof,
-		ChallengedSectors: sectors,
-		Prover:            abi.ActorID(mid),
-	})
-	if err != nil {
-		return xerrors.Errorf("failed to verify election post: %w", err)
-	}
-
-	if !ok {
-		log.Errorf("invalid winning post (block: %s, %x; %v)", h.Cid(), rand, sectors)
-		return xerrors.Errorf("winning post was invalid")
-	}
-
-	return nil
-}
-
-// TODO: We should extract this somewhere else and make the message pool and miner use the same logic
-func (syncer *Syncer) checkBlockMessages(ctx context.Context, b *types.FullBlock, baseTs *types.TipSet) error {
-	{
-		var sigCids []cid.Cid // this is what we get for people not wanting the marshalcbor method on the cid type
-		var pubks [][]byte
-
-		for _, m := range b.BlsMessages {
-			sigCids = append(sigCids, m.Cid())
-
-			pubk, err := syncer.sm.GetBlsPublicKey(ctx, m.From, baseTs)
-			if err != nil {
-				return xerrors.Errorf("failed to load bls public to validate block: %w", err)
-			}
-
-			pubks = append(pubks, pubk)
-		}
-
-		if err := syncer.verifyBlsAggregate(ctx, b.Header.BLSAggregate, sigCids, pubks); err != nil {
-			return xerrors.Errorf("bls aggregate signature was invalid: %w", err)
-		}
-	}
-
-	nonces := make(map[address.Address]uint64)
-
-	stateroot, _, err := syncer.sm.TipSetState(ctx, baseTs)
-	if err != nil {
-		return err
-	}
-
-	st, err := state.LoadStateTree(syncer.store.ActorStore(ctx), stateroot)
-	if err != nil {
-		return xerrors.Errorf("failed to load base state tree: %w", err)
-	}
-
-	nv := syncer.sm.GetNtwkVersion(ctx, b.Header.Height)
-	pl := vm.PricelistByEpoch(baseTs.Height())
-	var sumGasLimit int64
-	checkMsg := func(msg types.ChainMsg) error {
-		m := msg.VMMessage()
-
-		// Phase 1: syntactic validation, as defined in the spec
-		minGas := pl.OnChainMessage(msg.ChainLength())
-		if err := m.ValidForBlockInclusion(minGas.Total(), nv); err != nil {
-			return err
-		}
-
-		// ValidForBlockInclusion checks if any single message does not exceed BlockGasLimit
-		// So below is overflow safe
-		sumGasLimit += m.GasLimit
-		if sumGasLimit > build.BlockGasLimit {
-			return xerrors.Errorf("block gas limit exceeded")
-		}
-
-		// Phase 2: (Partial) semantic validation:
-		// the sender exists and is an account actor, and the nonces make sense
-		var sender address.Address
-		if nv >= network.Version13 {
-			sender, err = st.LookupID(m.From)
-			if err != nil {
-				return err
-			}
-		} else {
-			sender = m.From
-		}
-
-		if _, ok := nonces[sender]; !ok {
-			// `GetActor` does not validate that this is an account actor.
-			act, err := st.GetActor(sender)
-			if err != nil {
-				return xerrors.Errorf("failed to get actor: %w", err)
-			}
-
-			if !builtin.IsAccountActor(act.Code) {
-				return xerrors.New("Sender must be an account actor")
-			}
-			nonces[sender] = act.Nonce
-		}
-
-		if nonces[sender] != m.Nonce {
-			return xerrors.Errorf("wrong nonce (exp: %d, got: %d)", nonces[sender], m.Nonce)
-		}
-		nonces[sender]++
-
-		return nil
-	}
-
-	// Validate message arrays in a temporary blockstore.
-	tmpbs := bstore.NewMemory()
-	tmpstore := blockadt.WrapStore(ctx, cbor.NewCborStore(tmpbs))
-
-	bmArr := blockadt.MakeEmptyArray(tmpstore)
-	for i, m := range b.BlsMessages {
-		if err := checkMsg(m); err != nil {
-			return xerrors.Errorf("block had invalid bls message at index %d: %w", i, err)
-		}
-
-		c, err := store.PutMessage(tmpbs, m)
-		if err != nil {
-			return xerrors.Errorf("failed to store message %s: %w", m.Cid(), err)
-		}
-
-		k := cbg.CborCid(c)
-		if err := bmArr.Set(uint64(i), &k); err != nil {
-			return xerrors.Errorf("failed to put bls message at index %d: %w", i, err)
-		}
-	}
-
-	smArr := blockadt.MakeEmptyArray(tmpstore)
-	for i, m := range b.SecpkMessages {
-		if err := checkMsg(m); err != nil {
-			return xerrors.Errorf("block had invalid secpk message at index %d: %w", i, err)
-		}
-
-		// `From` being an account actor is only validated inside the `vm.ResolveToKeyAddr` call
-		// in `StateManager.ResolveToKeyAddress` here (and not in `checkMsg`).
-		kaddr, err := syncer.sm.ResolveToKeyAddress(ctx, m.Message.From, baseTs)
-		if err != nil {
-			return xerrors.Errorf("failed to resolve key addr: %w", err)
-		}
-
-		if err := sigs.Verify(&m.Signature, kaddr, m.Message.Cid().Bytes()); err != nil {
-			return xerrors.Errorf("secpk message %s has invalid signature: %w", m.Cid(), err)
-		}
-
-		c, err := store.PutMessage(tmpbs, m)
-		if err != nil {
-			return xerrors.Errorf("failed to store message %s: %w", m.Cid(), err)
-		}
-		k := cbg.CborCid(c)
-		if err := smArr.Set(uint64(i), &k); err != nil {
-			return xerrors.Errorf("failed to put secpk message at index %d: %w", i, err)
-		}
-	}
-
-	bmroot, err := bmArr.Root()
-	if err != nil {
-		return err
-	}
-
-	smroot, err := smArr.Root()
-	if err != nil {
-		return err
-	}
-
-	mrcid, err := tmpstore.Put(ctx, &types.MsgMeta{
-		BlsMessages:   bmroot,
-		SecpkMessages: smroot,
-	})
-	if err != nil {
-		return err
-	}
-
-	if b.Header.Messages != mrcid {
-		return fmt.Errorf("messages didnt match message root in header")
-	}
-
-	// Finally, flush.
-	return vm.Copy(ctx, tmpbs, syncer.store.ChainBlockstore(), mrcid)
-}
-
-func (syncer *Syncer) verifyBlsAggregate(ctx context.Context, sig *crypto.Signature, msgs []cid.Cid, pubks [][]byte) error {
-	_, span := trace.StartSpan(ctx, "syncer.verifyBlsAggregate")
-	defer span.End()
-	span.AddAttributes(
-		trace.Int64Attribute("msgCount", int64(len(msgs))),
-	)
-
-	msgsS := make([]ffi.Message, len(msgs))
-	pubksS := make([]ffi.PublicKey, len(msgs))
-	for i := 0; i < len(msgs); i++ {
-		msgsS[i] = msgs[i].Bytes()
-		copy(pubksS[i][:], pubks[i][:ffi.PublicKeyBytes])
-	}
-
-	sigS := new(ffi.Signature)
-	copy(sigS[:], sig.Data[:ffi.SignatureBytes])
-
-	if len(msgs) == 0 {
-		return nil
-	}
-
-	valid := ffi.HashVerify(sigS, msgsS, pubksS)
-	if !valid {
-		return xerrors.New("bls aggregate signature failed to verify")
-	}
 	return nil
 }
 
@@ -1374,7 +801,7 @@ loop:
 			return nil, xerrors.Errorf("retrieved segments of the chain are not connected at heights %d/%d",
 				blockSet[len(blockSet)-1].Height(), blks[0].Height())
 			// A successful `GetBlocks()` call is guaranteed to fetch at least
-			// one tipset so the acess `blks[0]` is safe.
+			// one tipset so the access `blks[0]` is safe.
 		}
 
 		for _, b := range blks {
@@ -1598,6 +1025,35 @@ func (syncer *Syncer) iterFullTipsets(ctx context.Context, headers []*types.TipS
 	return nil
 }
 
+func checkMsgMeta(ts *types.TipSet, allbmsgs []*types.Message, allsmsgs []*types.SignedMessage, bmi, smi [][]uint64) error {
+	for bi, b := range ts.Blocks() {
+		if msgc := len(bmi[bi]) + len(smi[bi]); msgc > build.BlockMessageLimit {
+			return fmt.Errorf("block %q has too many messages (%d)", b.Cid(), msgc)
+		}
+
+		var smsgCids []cid.Cid
+		for _, m := range smi[bi] {
+			smsgCids = append(smsgCids, allsmsgs[m].Cid())
+		}
+
+		var bmsgCids []cid.Cid
+		for _, m := range bmi[bi] {
+			bmsgCids = append(bmsgCids, allbmsgs[m].Cid())
+		}
+
+		mrcid, err := computeMsgMeta(cbor.NewCborStore(bstore.NewMemory()), bmsgCids, smsgCids)
+		if err != nil {
+			return err
+		}
+
+		if b.Messages != mrcid {
+			return fmt.Errorf("messages didnt match message root in header for ts %s", ts.Key())
+		}
+	}
+
+	return nil
+}
+
 func (syncer *Syncer) fetchMessages(ctx context.Context, headers []*types.TipSet, startOffset int) ([]*exchange.CompactedMessages, error) {
 	batchSize := len(headers)
 	batch := make([]*exchange.CompactedMessages, batchSize)
@@ -1636,7 +1092,19 @@ func (syncer *Syncer) fetchMessages(ctx context.Context, headers []*types.TipSet
 					if err != nil {
 						requestErr = multierror.Append(requestErr, err)
 					} else {
-						requestResult = result
+						isGood := true
+						for index, ts := range headers[nextI:lastI] {
+							cm := result[index]
+							if err := checkMsgMeta(ts, cm.Bls, cm.Secpk, cm.BlsIncludes, cm.SecpkIncludes); err != nil {
+								log.Errorf("fetched messages not as expected: %s", err)
+								isGood = false
+								break
+							}
+						}
+
+						if isGood {
+							requestResult = result
+						}
 					}
 				}
 
@@ -1754,10 +1222,6 @@ func (syncer *Syncer) collectChain(ctx context.Context, ts *types.TipSet, hts *t
 	return nil
 }
 
-func VerifyElectionPoStVRF(ctx context.Context, worker address.Address, rand []byte, evrf []byte) error {
-	return gen.VerifyVRF(ctx, worker, rand, evrf)
-}
-
 func (syncer *Syncer) State() []SyncerStateSnapshot {
 	return syncer.syncmgr.State()
 }
@@ -1801,13 +1265,4 @@ func (syncer *Syncer) getLatestBeaconEntry(_ context.Context, ts *types.TipSet) 
 	}
 
 	return nil, xerrors.Errorf("found NO beacon entries in the 20 latest tipsets")
-}
-
-func (syncer *Syncer) IsEpochBeyondCurrMax(epoch abi.ChainEpoch) bool {
-	if syncer.Genesis == nil {
-		return false
-	}
-
-	now := uint64(build.Clock.Now().Unix())
-	return epoch > (abi.ChainEpoch((now-syncer.Genesis.MinTimestamp())/build.BlockDelaySecs) + MaxHeightDrift)
 }
