+++ itests/kit/ensemble_opts_nv.go
@@ -3,6 +3,8 @@ package kit
 import (
 	"github.com/filecoin-project/go-state-types/abi"
 	"github.com/filecoin-project/go-state-types/network"
+
+	"github.com/filecoin-project/lotus/chain/consensus/filcns"
 	"github.com/filecoin-project/lotus/chain/stmgr"
 )
 
@@ -28,7 +30,7 @@ func SDRUpgradeAt(calico, persian abi.ChainEpoch) EnsembleOpt {
 	}, stmgr.Upgrade{
 		Network:   network.Version7,
 		Height:    calico,
-		Migration: stmgr.UpgradeCalico,
+		Migration: filcns.UpgradeCalico,
 	}, stmgr.Upgrade{
 		Network: network.Version8,
 		Height:  persian,
@@ -42,7 +44,7 @@ func LatestActorsAt(upgradeHeight abi.ChainEpoch) EnsembleOpt {
 	}, stmgr.Upgrade{
 		Network:   network.Version13,
 		Height:    upgradeHeight,
-		Migration: stmgr.UpgradeActorsV5,
+		Migration: filcns.UpgradeActorsV5,
 	})
 }
 
@@ -53,6 +55,6 @@ func TurboUpgradeAt(upgradeHeight abi.ChainEpoch) EnsembleOpt {
 	}, stmgr.Upgrade{
 		Network:   network.Version12,
 		Height:    upgradeHeight,
-		Migration: stmgr.UpgradeActorsV4,
+		Migration: filcns.UpgradeActorsV4,
 	})
 }
