+++ node/impl/full/chain.go
@@ -11,6 +11,7 @@ import (
 	"sync"
 
 	"github.com/filecoin-project/lotus/build"
+	"github.com/filecoin-project/lotus/chain/stmgr"
 
 	"go.uber.org/fx"
 	"golang.org/x/xerrors"
@@ -78,7 +79,8 @@ type ChainAPI struct {
 	WalletAPI
 	ChainModuleAPI
 
-	Chain *store.ChainStore
+	Chain  *store.ChainStore
+	TsExec stmgr.Executor
 
 	// ExposedBlockstore is the global monolith blockstore that is safe to
 	// expose externally. In the future, this will be segregated into two
@@ -394,7 +396,7 @@ func (s stringKey) Key() string {
 
 // TODO: ActorUpgrade: this entire function is a problem (in theory) as we don't know the HAMT version.
 // In practice, hamt v0 should work "just fine" for reading.
-func resolveOnce(bs blockstore.Blockstore) func(ctx context.Context, ds ipld.NodeGetter, nd ipld.Node, names []string) (*ipld.Link, []string, error) {
+func resolveOnce(bs blockstore.Blockstore, tse stmgr.Executor) func(ctx context.Context, ds ipld.NodeGetter, nd ipld.Node, names []string) (*ipld.Link, []string, error) {
 	return func(ctx context.Context, ds ipld.NodeGetter, nd ipld.Node, names []string) (*ipld.Link, []string, error) {
 		store := adt.WrapStore(ctx, cbor.NewCborStore(bs))
 
@@ -468,7 +470,7 @@ func resolveOnce(bs blockstore.Blockstore) func(ctx context.Context, ds ipld.Nod
 				}, nil, nil
 			}
 
-			return resolveOnce(bs)(ctx, ds, n, names[1:])
+			return resolveOnce(bs, tse)(ctx, ds, n, names[1:])
 		}
 
 		if strings.HasPrefix(names[0], "@A:") {
@@ -517,7 +519,7 @@ func resolveOnce(bs blockstore.Blockstore) func(ctx context.Context, ds ipld.Nod
 				}, nil, nil
 			}
 
-			return resolveOnce(bs)(ctx, ds, n, names[1:])
+			return resolveOnce(bs, tse)(ctx, ds, n, names[1:])
 		}
 
 		if names[0] == "@state" {
@@ -531,7 +533,7 @@ func resolveOnce(bs blockstore.Blockstore) func(ctx context.Context, ds ipld.Nod
 				return nil, nil, xerrors.Errorf("getting actor head for @state: %w", err)
 			}
 
-			m, err := vm.DumpActorState(&act, head.RawData())
+			m, err := vm.DumpActorState(tse.NewActorRegistry(), &act, head.RawData())
 			if err != nil {
 				return nil, nil, err
 			}
@@ -565,7 +567,7 @@ func resolveOnce(bs blockstore.Blockstore) func(ctx context.Context, ds ipld.Nod
 				}, nil, nil
 			}
 
-			return resolveOnce(bs)(ctx, ds, n, names[1:])
+			return resolveOnce(bs, tse)(ctx, ds, n, names[1:])
 		}
 
 		return nd.ResolveLink(names)
@@ -585,7 +587,7 @@ func (a *ChainAPI) ChainGetNode(ctx context.Context, p string) (*api.IpldObject,
 
 	r := &resolver.Resolver{
 		DAG:         dag,
-		ResolveOnce: resolveOnce(bs),
+		ResolveOnce: resolveOnce(bs, a.TsExec),
 	}
 
 	node, err := r.ResolvePath(ctx, ip)
